refactor: remove redundant "page" suffix from file names in pages folder
